requirements.txt
absl-py==0.5.0
astor==0.7.1
blessings==1.7
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
inquirer==2.5.1
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==2.6.11
numpy==1.14.5
pkg-resources==0.0.0
protobuf==3.6.1
PyYAML==3.13
readchar==2.0.1
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
textgenrnn==1.4
Werkzeug==0.14.1
